# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Roombnb API",
        "description": "API para listagem de rooms, criação de reservas e listagem de reservas.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Ambiente local"
        }
    ],
    "paths": {
        "/api/v1/reservations": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Lista reservas de um usuário",
                "description": "Retorna todas as reservas associadas ao user_id informado.",
                "operationId": "1fea0748719a5cbced2b611e67240778",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "ID do usuário (tenant) dono das reservas",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 4
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de reservas do usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 15
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "pending"
                                                    },
                                                    "payment_method": {
                                                        "type": "string",
                                                        "example": "pix",
                                                        "nullable": true
                                                    },
                                                    "check_in": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2025-02-11"
                                                    },
                                                    "check_out": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2025-02-12"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-01-10T12:34:56"
                                                    },
                                                    "room": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 7
                                                            },
                                                            "title": {
                                                                "type": "string",
                                                                "example": "Consultório no Centro"
                                                            },
                                                            "city": {
                                                                "properties": {
                                                                    "name": {
                                                                        "type": "string",
                                                                        "example": "Salvador"
                                                                    },
                                                                    "state": {
                                                                        "type": "string",
                                                                        "example": "BA"
                                                                    }
                                                                },
                                                                "type": "object",
                                                                "nullable": true
                                                            },
                                                            "specialty": {
                                                                "properties": {
                                                                    "name": {
                                                                        "type": "string",
                                                                        "example": "Psicologia"
                                                                    }
                                                                },
                                                                "type": "object",
                                                                "nullable": true
                                                            }
                                                        },
                                                        "type": "object",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Parâmetro user_id ausente ou inválido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "user_id is required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/rooms/{room}/reservations": {
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Cria uma nova reserva para um room",
                "operationId": "e384f7a737c1b5d62001855156329c37",
                "parameters": [
                    {
                        "name": "room",
                        "in": "path",
                        "description": "ID do room",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 7
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "check_in",
                                    "check_out"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "check_in": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-02-11"
                                    },
                                    "check_out": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-02-12"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "credit_card",
                                            "cash",
                                            "pix"
                                        ],
                                        "example": "pix"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reserva criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                },
                                                "payment_method": {
                                                    "type": "string",
                                                    "example": "pix"
                                                },
                                                "check_in": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2025-02-11"
                                                },
                                                "check_out": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2025-02-12"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Reserva criada com sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação ou conflito de datas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Os dados enviados são inválidos."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Token JWT ou Sanctum no formato: Bearer {token}"
            }
        }
    },
...
}